+++ server/src/users/users.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { User } from './users.model';
 import { InjectModel } from '@nestjs/mongoose';
 import mongoose, { Model } from 'mongoose';
-import { UsersRepository } from './user.repository';
+import { UsersRepository } from './users.repository';
 @Injectable()
 export class UsersService {
     
@@ -18,8 +18,7 @@ export class UsersService {
             if(!id){
                 return null;
             }
-            const user = await this.usersRepository.findUserById(id);
-            return user;
+            return await this.usersRepository.findUserById(id);;
         } catch (error) {
             throw new Error(error);
         }
@@ -29,114 +28,68 @@ export class UsersService {
             if(!id){
                 return null;
             }
-            const user = await this.userModel.findOne({_id: id});
-            return user;
+            return await this.usersRepository.findUserByObjectId(id);;
         } catch (error) {
             throw new Error(error);
         }
     }
-    async findOne(email: string){
-        return this.userModel.find(user => user.email === email);
-    }
-    async findByEmail(email: string){
+    async findUserByEmail(email: string){
         try {
-            return this.userModel.findOne({email: email});
+            return await this.usersRepository.findByEmail(email);
         } catch (error) {
             throw new Error(error);
         }
     }
     async addChatGroupToUser(userId:string, chatGroup:object){
         try {
-            const updatedUser = await this.userModel.findByIdAndUpdate(
-                userId,
-                {$push:{ChatGroups:chatGroup}},
-                {new:true}
-            );
-            await updatedUser.save();
-            return updatedUser;
+            return await this.usersRepository.addChatGroupToUser(userId, chatGroup);;
         } catch (error) {
             throw new Error(error);
         }
     }
     async removeChatGroupFromUser(user , chatGroup ){
         try {
-            const userId: string = user._id;
-            const chatGroupId: string = chatGroup._id;
-
-            const updatedUser = await this.userModel.findByIdAndUpdate(
-                userId,
-                { $pull: { ChatGroups: {_id: chatGroupId } } },
-                { new:true }
-            );
-            await updatedUser.save();
+            await this.usersRepository.removeChatGroupFromUser(user,chatGroup);
         } catch (error) {
             throw new Error(error);
         }
     }
     async addFriend(userId:string, friend:object) {
         try {
-            const updatedUser = await this.userModel.findByIdAndUpdate(
-                userId,
-                {$push:{Friends:friend}},
-                {new:true}
-            );
-            await updatedUser.save();
-            return updatedUser;
+            return await this.usersRepository.addFriend(userId, friend);;
         } catch (error) {
             throw new Error(error);
         }
     }
     async removeFriend(userId:string, friendId:string) {
         try {
-            const updatedUser = await this.userModel.findByIdAndUpdate(
-                userId,
-                { $pull: { Friends: { _id: friendId } } },
-                { new: true }
-            );
-            await updatedUser.save();
-            return updatedUser;
+            return await this.usersRepository.removeFriend(userId, friendId);
         } catch (error) {
             throw new Error(error);
         }
     }
     async getFriendsOfUser( userId:mongoose.Types.ObjectId) {
         try {
-            const user = await this.findUser(userId);
-            const friends : object[] = user.Friends;
-            return friends;
+            return await this.usersRepository.getFriendsOfUser(userId);
         } catch (error) {
             throw new Error(error);
         }
     }
     async getFriendsOfUserById( userId:string) {
         try {
-            const user = await this.findUserById(userId);
-            const friends : object[] = user.Friends;
-            return friends;
+            return await this.usersRepository.getFriendsOfUserById(userId);
         } catch (error) {
             throw new Error(error);
         }
     }
     async getUserData( userObject: Object ){
         try {
-            const user = await this.userModel.findOne(userObject);
-            const { name, email, ChatGroups, _id } = user;
-            const userData = {
-                _id: _id,
-                name: name,
-                email: email,
-                ChatGroups: ChatGroups
-            }
-            return userData;
+            return await this.usersRepository.getUserData(userObject);
         } catch (error) {
             throw new Error(error);
         }
     }
     async searchUser(searchText:string) {
-        const users = await this.userModel.find({
-            $or: [{ name: { $regex: searchText, $options: "i" } }, { email: { $regex: searchText, $options: "i" } }],
-        })
-        .exec();
-        return users;
+        return await this.usersRepository.searchUser(searchText);
     }
 }
